Changed shooting classes and health class

+++ CM/Scripts/CM/Essentials/IDamageable.cs
@@ -2,6 +2,6 @@
 {
 	public interface IDamageable
 	{
-		void TakeDamage(int damageValue);
+		void TakeDamage(float damageValue);
 	}
 }
